refactor: move menu-item to under nav
